+++ test/integration/case-projects/args/test-config.js
@@ -1,0 +1,4 @@
+module.exports = {
+  args: ['Main.elm', 'hello', 'world'],
+  expectedStdout: 'hello-world!\n',
+};