+++ test/integration/case-projects/custom-constant/test-config.js
@@ -1,0 +1,4 @@
+module.exports = {
+  args: ['--output-name=customConstant', 'Main.elm'],
+  expectedStdout: 'static output from custom constant\n',
+};