+++ test/integration/case-projects/custom-function/test-config.js
@@ -1,0 +1,4 @@
+module.exports = {
+  args: ['--output-name=customFunction', 'Main.elm', 'forty', 'two'],
+  expectedStdout: 'static output from forty two\n',
+};