+++ test/integration/case-projects/long-output/test-config.js
@@ -1,0 +1,4 @@
+module.exports = {
+  args: ['Main.elm'],
+  expectedStdout: () => 'static output\n'.repeat(1000000),
+};